requirements.txt
streamlit
scikit-learn
pandas
numpy
sentence-transformers
faiss-cpu
